pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Gen diff"
author = "Me"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml==6.0.2",
    "ruff>=0.9.3",
    "uv==0.5.7",
    "yml==0.0.1",
]

[project.scripts]
gendiff = "gendiff.scripts.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[dependency-groups]
dev = [
    "coverage>=7.6.11",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
]
